backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
libsql-client>=0.3.0
python-dotenv>=1.0.0
httpx>=0.26.0
google-auth>=2.27.0
PyJWT>=2.8.0
